Reorganize filtering and sorting components into a subfolder
