+++ ui/src/SortingAndFiltering/NewFilter.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Application/Styleguide/Colors.scss';
+@import '../Application/Styleguide/Colors';
 
 .newFilterContainer {
   display: flex;
