+++ ui/src/SortingAndFiltering/NewFilterOrSortBy.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Application/Styleguide/Colors.scss';
+@import '../Application/Styleguide/Colors';
 
 .newDropdownContainer {
   display: flex;
