+++ ui/src/SortingAndFiltering/ProductFilterOrSortBy.scss
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-@import '../Application/Styleguide/Colors.scss';
-@import '../Application/Styleguide/Breakpoints.scss';
+@import '../Application/Styleguide/Colors';
+@import '../Application/Styleguide/Breakpoints';
 
 .dropdown {
   width: auto;
