+++ ui/src/tests/LocalStorageFilters.test.tsx
@@ -19,7 +19,7 @@ import TestUtils, {renderWithRedux} from './TestUtils';
 import PeopleMover from '../Application/PeopleMover';
 import {findByText, fireEvent} from '@testing-library/dom';
 import React from 'react';
-import {LocalStorageFilters} from '../ReusableComponents/ProductFilter';
+import {LocalStorageFilters} from '../SortingAndFiltering/ProductFilter';
 import selectEvent from 'react-select-event';
 import {RenderResult, wait} from '@testing-library/react';
 import {createBrowserHistory} from 'history';
